[ACP-164] removed the `route` property for now
I'm ready to accept the pull request, without the `route` property. This is not because of a general objection against `route`, but a confusion what the role of `route` is. In particular, the current description as suggested in the schema seems to expose too much unnecessary and unexplained implementation detail like:
```json
                            "description": "The SKMS data center that the message orginated from."
```

The "SKMS" acronym is not explained elsewhere in XDM. 

In order to get the pull request accepted quickly, I'm merging with `route` excluded. Please open a new PR that sufficiently documents and explains `route`