Added requirement that XDM properties always have concrete types.
It was found that with enumeration support that types can be variable. This is something that can cause great difficulties in data interoperability and code generation from JSON Schema. So to close this hole I am proposing that all properties have concrete types. 

Example of what we do not want to allow is :
Even more than that: the enum constraint can go without accompanying type constraint and contain values of different type, e.g.
{
    "$schema": "http://json-schema.org/draft-06/schema#",
    "$id": "https://ns.adobe.com/xdm/sample1", 
    "definitions": {
      "False": {
        "enum": [0, false, "false"]
      } 
    },
    "$ref": "#/definitions/False"
}