Removing dependency of principal from user schema. This ensures that mandatory similar properties of principal do not become mandatory properties for user schema.
I think the existing user.example.1.json example makes it seem that “principal” properties in user.schema.json were supposed to be optional. For example, @id is optional.
This is the current user.example.1.json
{
  "xdm:identityProvider": "ims",
  "xdm:id": "C0B648DE57D701277F000101@AdobeID",
  "xdm:type": "https://ns.adobe.com/adobecloudplatform/ims/user"
}
XDM registry rejects this example and expects a XDM object that matches user.schema.json to be like -
{
  "xdm:identityProvider": "ims",
  "xdm:id": "C0B648DE57D701277F000101@AdobeID",
  "xdm:type": "https://ns.adobe.com/adobecloudplatform/ims/user"
  "xdm:provider": {
    "@id": "https://ims-na1.adobelogin.com/"
  },
  "@id": "C0B648DE57D701277F000101@AdobeID",
  "@type": "https://ns.adobe.com/adobecloudplatform/ims/user"
}
This looks really awkward to have the similar values duplicated twice in xdm:id and @id. Similarly, @type and xdm:type. I think @type is not the preferred option compared to xdm:type as the latter has a namespace. This pull request drops import of principal.schema.json in user.schema.json. This would ensure that user.example.1.json passes XDM registry test for user.schema.json.