Add xdm:note to the alternateDisplayInfo descriptor
AEP PM is wanting to introduce a way for users to add additional user specific "notes" to a given field (how it should be used, etc.). They see this as something different from the description and would like it to be separate. Adding this attribute to the descriptor will enable a user to add a note to one of the standard fields.