Adding min & max for ProcessingTimeMS
ProcessingTimeMS is of type int and meta:xdmType as long. Since it does not have minimum and maximum defined, it uses int type's min and max. In order for it to have min and max of Long, it have to be explicitly added.  Reference - https://experienceleague.adobe.com/docs/experience-platform/xdm/schema/field-constraints.html?lang=en